// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/models/ClubSearchResult.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClubSearchResult = {
    clubId: number | string;
    clubName: string;
    state?: string | null;
    teamCount: number | string;
    matchScore: number | string;
};